fix(pci-instance): rework zod-form/ComboboxField
ref: #TAPC-5116
